+++ tests/misc_test.py
@@ -436,6 +436,26 @@ class AuxDataTests(unittest.TestCase):
 
             self.assertEqual(m.aux_data["elfSoname"].data, binary)
 
+    @unittest.skipUnless(
+        platform.system() == "Linux", "This test is linux only."
+    )
+    def test_aligned_data_in_code(self):
+        """
+        Test that alignment directives are correctly generated for
+        data_in_code referenced by instructions that require aligned memory.
+        """
+        binary = "ex"
+        with cd(ex_asm_dir / "ex_aligned_data_in_code"):
+            self.assertTrue(compile("gcc", "g++", "-O0", []))
+            ir = disassemble(Path(binary)).ir()
+            m = ir.modules[0]
+
+            alignments = m.aux_data["alignment"].data.items()
+            alignment_list = [alignment for uuid, alignment in alignments]
+            self.assertEqual(alignment_list.count(16), 2)
+            self.assertEqual(alignment_list.count(32), 1)
+            self.assertEqual(alignment_list.count(64), 2)
+
 
 class RawGtirbTests(unittest.TestCase):
     @unittest.skipUnless(